Localization based on both IR sides is now working. Further the merging of walls has been improved. It is now possible to correct even wrong position estimations which would result in the robot beeing inside of a previously mapped wall. The Localization has still some issues though: At wall edges it appears as if it the localization is wrong about theta. This happens probably due to the weird behavior of the IRs at wall edges.
If the pose estimation is off, it maps walls anyway. A correction short after mapping a wall
corrects the pose, however, not the wall that has been mapped with the
incorrect pose. This can result in an incorrect map! Hopefully this won't be
a problem once a Kalman filter is used.